+++ jnosql-mapping/jnosql-mapping-core/src/test/java/org/eclipse/jnosql/mapping/core/util/ConverterUtilTest.java
@@ -19,7 +19,7 @@ import org.eclipse.jnosql.mapping.core.entities.Money;
 import org.eclipse.jnosql.mapping.core.entities.Person;
 import org.eclipse.jnosql.mapping.core.entities.Worker;
 import org.eclipse.jnosql.mapping.reflection.Reflections;
-import org.eclipse.jnosql.mapping.core.spi.ReflectionEntityMetadataExtension;
+import org.eclipse.jnosql.mapping.reflection.spi.ReflectionEntityMetadataExtension;
 import org.jboss.weld.junit5.auto.AddExtensions;
 import org.jboss.weld.junit5.auto.AddPackages;
 import org.jboss.weld.junit5.auto.EnableAutoWeld;
