+++ jnosql-mapping/jnosql-mapping-reflection/src/main/resources/META-INF/services/jakarta.enterprise.inject.spi.Extension
@@ -13,4 +13,4 @@
 #   Otavio Santana
 #
 
-org.eclipse.jnosql.mapping.core.spi.ReflectionEntityMetadataExtension
+org.eclipse.jnosql.mapping.reflection.spi.ReflectionEntityMetadataExtension
