+++ jnosql-mapping/jnosql-mapping-semistructured/src/test/java/org/eclipse/jnosql/mapping/semistructured/EntityConverterInheritanceTest.java
@@ -30,7 +30,7 @@ import org.eclipse.jnosql.mapping.semistructured.entities.inheritance.SmallProje
 import org.eclipse.jnosql.mapping.semistructured.entities.inheritance.SmsNotification;
 import org.eclipse.jnosql.mapping.semistructured.entities.inheritance.SocialMediaNotification;
 import org.eclipse.jnosql.mapping.reflection.Reflections;
-import org.eclipse.jnosql.mapping.core.spi.ReflectionEntityMetadataExtension;
+import org.eclipse.jnosql.mapping.reflection.spi.ReflectionEntityMetadataExtension;
 import org.jboss.weld.junit5.auto.AddExtensions;
 import org.jboss.weld.junit5.auto.AddPackages;
 import org.jboss.weld.junit5.auto.EnableAutoWeld;
